requirements.txt
# Lighthouse Trading — Python dependencies
# Install: pip install -r requirements.txt

# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Configuration
python-dotenv>=1.0.0

# Data validation (used by FastAPI)
pydantic>=2.0.0

# Exchange SDKs
hyperliquid-python-sdk>=0.9.0
python-binance>=1.0.19

# Ethereum (required by hyperliquid SDK)
eth-account>=0.11.0

# HTTP client (for Telegram)
requests>=2.31.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0       # async test client for FastAPI

# Typing
typing-extensions>=4.9.0
